backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
strawberry-graphql[fastapi]==0.245.0
python-dotenv==1.0.1
pymongo==4.10.1
pydantic==2.9.2
python-multipart==0.0.17
httpx==0.27.2
tensorflow==2.20.0
pillow==11.0.0
numpy==2.2.0
scikit-learn==1.5.2
pandas==2.2.3
joblib==1.4.2
